refactor: split CLI into core and command modules
