+++ yonote_cli/yonote_cli/commands/__init__.py
@@ -1,0 +1,28 @@
+"""Command handlers for yonote CLI."""
+
+from .auth import cmd_auth_set, cmd_auth_info
+from .cache import cache_info, cache_clear
+from .collections import cmd_collections_list, cmd_collections_export
+from .documents import (
+    cmd_docs_list,
+    cmd_docs_export,
+    cmd_docs_export_batch,
+    cmd_docs_import,
+    cmd_docs_import_dir,
+    cmd_docs_tree,
+)
+
+__all__ = [
+    "cmd_auth_set",
+    "cmd_auth_info",
+    "cache_info",
+    "cache_clear",
+    "cmd_collections_list",
+    "cmd_collections_export",
+    "cmd_docs_list",
+    "cmd_docs_export",
+    "cmd_docs_export_batch",
+    "cmd_docs_import",
+    "cmd_docs_import_dir",
+    "cmd_docs_tree",
+]
